add type convert
