+++ src/fsrs/help.ts
@@ -1,6 +1,7 @@
 import type { int, unit } from './types'
 import type { DateInput, Grade } from './models'
 import { Rating, State } from './models'
+import { TypeConvert } from './convert'
 
 declare global {
   export interface Date {
@@ -120,50 +121,26 @@ export function show_diff_message(
   return `${Math.floor(diff)}${unit ? timeUnit[i] : ''}`
 }
 
+/**
+ *
+ * @deprecated Use TypeConvert.time instead
+ */
 export function fixDate(value: unknown) {
-  if (typeof value === 'object' && value instanceof Date) {
-    return value
-  } else if (typeof value === 'string') {
-    const timestamp = Date.parse(value)
-    if (!isNaN(timestamp)) {
-      return new Date(timestamp)
-    } else {
-      throw new Error(`Invalid date:[${value}]`)
-    }
-  } else if (typeof value === 'number') {
-    return new Date(value)
-  }
-  throw new Error(`Invalid date:[${value}]`)
+  return TypeConvert.time(value)
 }
 
+/**
+ * @deprecated Use TypeConvert.state instead
+ */
 export function fixState(value: unknown): State {
-  if (typeof value === 'string') {
-    const firstLetter = value.charAt(0).toUpperCase()
-    const restOfString = value.slice(1).toLowerCase()
-    const ret = State[`${firstLetter}${restOfString}` as keyof typeof State]
-    if (ret === undefined) {
-      throw new Error(`Invalid state:[${value}]`)
-    }
-    return ret
-  } else if (typeof value === 'number') {
-    return value as State
-  }
-  throw new Error(`Invalid state:[${value}]`)
+  return TypeConvert.state(value)
 }
 
+/**
+ * @deprecated Use TypeConvert.rating instead
+ */
 export function fixRating(value: unknown): Rating {
-  if (typeof value === 'string') {
-    const firstLetter = value.charAt(0).toUpperCase()
-    const restOfString = value.slice(1).toLowerCase()
-    const ret = Rating[`${firstLetter}${restOfString}` as keyof typeof Rating]
-    if (ret === undefined) {
-      throw new Error(`Invalid rating:[${value}]`)
-    }
-    return ret
-  } else if (typeof value === 'number') {
-    return value as Rating
-  }
-  throw new Error(`Invalid rating:[${value}]`)
+  return TypeConvert.rating(value)
 }
 
 export const Grades: Readonly<Grade[]> = [
