+++ src/fsrs/types.ts
@@ -1,7 +1,5 @@
 import type {
   Card,
-  CardInput,
-  DateInput,
   Grade,
   RecordLog,
   RecordLogItem,
@@ -12,7 +10,7 @@ export type int = number & { __int__: void }
 export type double = number & { __double__: void }
 
 export interface IScheduler {
-  init(card: Card, now: Date): this
+  new (card: Card, now: Date): this
   preview(): RecordLog
   review(state: Grade): RecordLogItem
   addPlugin(plugin: ISchedulerLifecycle): this
@@ -26,11 +24,4 @@ export interface ISchedulerLifecycle {
   update_state(): void
   schedule(): void
   display(): void
-}
-
-export interface ITransformer {
-  card<T extends Card | CardInput>(card: T): Card
-  recordLog<T extends RecordLog>(recordLog: T): RecordLog
-  recordLogItem<T extends RecordLogItem>(recordLogItem: T): RecordLogItem
-  time<T extends Date | DateInput>(time: T): Date
 }
